pyproject.toml
[project]
name = "sourcebox-articles"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "python-dotenv>=1.2.1",
    "replicate>=1.0.7",
    "requests>=2.32.5",
]
